+++ package.json
@@ -11,6 +11,7 @@
     "@testing-library/user-event": "^7.1.2",
     "animate.css": "^3.7.2",
     "peerjs": "^1.1.0",
+    "prop-types": "^15.7.2",
     "react": "^16.12.0",
     "react-animated-css": "^1.2.1",
     "react-dom": "^16.12.0",
